Create ClaimResponseFactory
